Add lifecycle & text API wrappers.
