+++ whycpp_c/include/whycpp_c/import_sprites.h
@@ -5,11 +5,14 @@
 #ifndef WHYCPP_ROOT_IMPORT_SPRITES_H
 #define WHYCPP_ROOT_IMPORT_SPRITES_H
 
+#include "c_api.h"
+
 /**
  * Loads sprite into memory and return sprite id
  * @param filename
  * @return
  */
+WHYCPP_C_API
 int ImportSprite_C(const char* filename);
 
-#endif //WHYCPP_ROOT_IMPORT_SPRITES_H
+#endif  // WHYCPP_ROOT_IMPORT_SPRITES_H
