+++ whycpp_c/src/lifecycle.cpp
@@ -1,0 +1,20 @@
+//
+// Created by k.leyfer on 29.03.2019.
+//
+
+#include <whycpp/lifecycle.h>
+#include <whycpp_c/lifecycle.h>
+
+#include "internal/global_context_int.h"
+
+void PauseApp_C() {
+  PauseApp(*GetContext_C());
+}
+
+void ResumeApp_C() {
+  ResumeApp(*GetContext_C());
+}
+
+void ExitApp_C() {
+  ExitApp(*GetContext_C());
+}
